LayoutTests/canvas/philip/tests/2d.imageData.create2.negative-expected.txt

< [index] >
2d.imageData.create2.negative
createImageData(sw, sh) takes the absolute magnitude of the size arguments
References: 2d.imageData.create2.size Bugzilla
Actual output:
Passed
